move Theme constants into ToolTheme object
