--- module/renderer/src/commonMain/kotlin/org/cru/godtools/shared/renderer/Theme.kt
@@ -1,9 +1,0 @@
-package org.cru.godtools.shared.renderer
-
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-
-internal val GodToolsGreen = Color(red = 0x6E, green = 0xDC, blue = 0x50)
-
-internal val ProgressBarHeight = 8.dp
-internal val ProgressBarGapSize = 0.dp - ProgressBarHeight
