+++ module/renderer/src/commonMain/kotlin/org/cru/godtools/shared/renderer/ToolTheme.kt
@@ -2,9 +2,12 @@ package org.cru.godtools.shared.renderer
 
 import androidx.compose.material3.CardDefaults
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 
 internal object ToolTheme {
+    internal val GodToolsGreen = Color(red = 0x6E, green = 0xDC, blue = 0x50)
+
     internal val ContentHorizontalPadding = 16.dp
 
     @Composable
@@ -16,4 +19,7 @@ internal object ToolTheme {
         disabledElevation = 3.dp,
     )
     internal val CardPadding = 8.dp
+
+    internal val ProgressBarHeight = 8.dp
+    internal val ProgressBarGapSize = 0.dp - ProgressBarHeight
 }
