+++ module/renderer/src/commonMain/kotlin/org/cru/godtools/shared/renderer/lesson/RenderLesson.kt
@@ -39,16 +39,15 @@ import com.github.ajalt.colormath.extensions.android.composecolor.toComposeColor
 import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.launch
 import org.ccci.gto.android.common.androidx.lifecycle.ConstrainedStateLifecycleOwner
-import org.cru.godtools.shared.renderer.ProgressBarGapSize
-import org.cru.godtools.shared.renderer.ProgressBarHeight
 import org.cru.godtools.shared.renderer.RenderBackground
+import org.cru.godtools.shared.renderer.ToolTheme.ProgressBarGapSize
+import org.cru.godtools.shared.renderer.ToolTheme.ProgressBarHeight
 import org.cru.godtools.shared.renderer.common.ToolLoading
 import org.cru.godtools.shared.renderer.common.ToolNotFound
 import org.cru.godtools.shared.renderer.common.ToolOffline
 import org.cru.godtools.shared.renderer.generated.resources.Res
 import org.cru.godtools.shared.renderer.generated.resources.lesson_accessibility_action_close
 import org.cru.godtools.shared.renderer.util.ContentEventListener
-import org.cru.godtools.shared.tool.parser.model.Manifest
 import org.cru.godtools.shared.tool.parser.model.lesson.LessonPage
 import org.cru.godtools.shared.tool.parser.model.lessonNavBarColor
 import org.cru.godtools.shared.tool.parser.model.lessonNavBarControlColor
